Move nix version check to its own module
Also declare min version in cargo config file
