+++ flake-info/.cargo/config.toml
@@ -1,0 +1,2 @@
+[env]
+MIN_NIX_VERSION = "2.7.0" # we need PRs #5878 and #5922 for package outputs
