+++ flake-info/src/commands/mod.rs
@@ -1,8 +1,10 @@
 mod nix_flake_attrs;
 mod nix_flake_info;
+mod nix_check_version;
 mod nix_gc;
 mod nixpkgs_info;
 pub use nix_flake_attrs::get_derivation_info;
 pub use nix_flake_info::get_flake_info;
+pub use nix_check_version::{check_nix_version, NixCheckError};
 pub use nix_gc::run_gc;
 pub use nixpkgs_info::{get_nixpkgs_info, get_nixpkgs_options};
